+++ spec/lib/danger/ci_sources/ci_source_spec.rb
@@ -21,6 +21,7 @@ RSpec.describe Danger::CI do
           "Danger::Codemagic",
           "Danger::Codeship",
           "Danger::Concourse",
+          "Danger::CustomCIWithGithub",
           "Danger::DotCi",
           "Danger::Drone",
           "Danger::GitHubActions",